% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_commas.R
\name{add_commas}
\alias{add_commas}
\title{Format numbers to add commas to thousands}
\usage{
add_commas(x, ...)
}
\arguments{
\item{x}{A list of numbers to be formatted}

\item{...}{Any additional arguments}
}
\value{
A formatted list of numbers
}
\description{
Format numbers to add commas to thousands
}
\examples{
x <- c(1000, 10000, 20000, 100, 4000)
add_commas(x)
}
\keyword{add_commas}
