% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scw_colours.R
\name{nhs_cols}
\alias{nhs_cols}
\title{Function to extract NHS colours as hex codes}
\usage{
nhs_cols(...)
}
\arguments{
\item{...}{Character names of NHS colours (e.g. "nhs_white")}
}
\description{
Function to extract NHS colours as hex codes
}
