% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-diverging.R
\name{pal_diverging}
\alias{pal_diverging}
\title{Diverging Colour Palettes}
\usage{
pal_diverging(
  palette = c("blue_green", "blue_yellow", "blue_red"),
  alpha = 1,
  reverse = FALSE
)
}
\arguments{
\item{palette}{Currently there are two diverging colour palettes available:
\code{"blue_green"} (a 9-colour blue-grey-green palette using NHS Blue, NHS
Pale Grey, and NHS Aqua Green), \code{blue_yellow_red} (a 9-colour
blue-yellow-red palette using NHS Blue and NHS Pale Grey), and
\code{blue_red} (a 9-colour blue-to-red palette uses NHS Blue and NHS Pale
Grey).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{The default, FALSE, generates the colour palette as defined,
while TRUE generates the colour palette in reverse order.}
}
\description{
Diverging colour palettes are used to encode numerical information that is
ordered and has a meaningful midpoint and meaningful extreme values. For
example, a five-point scale used to measure attitudes in a survey, from
Strongly Agree to Strongly Disagree, with Neither Agree Nor Disagree, would
require a diverging colour palette. Diverging palettes combine two different
hues, with the highest chroma (the intensity of a colour) representing the
extreme values, and the neutral midpoint value being the intersection of the
two hues.
}
\examples{
scales::show_col(pal_diverging(palette = "blue_green")(9))
scales::show_col(pal_diverging(palette = "blue_yellow")(9))
scales::show_col(pal_diverging(palette = "blue_red")(9))
scales::show_col(pal_diverging(palette = "blue_yellow", alpha = 0.7)(9))
scales::show_col(pal_diverging(palette = "blue_red", reverse = TRUE)(9))
}
