% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sequential.R
\name{pal_ramp}
\alias{pal_ramp}
\title{Generate Colour Palette}
\usage{
pal_ramp(values, discrete = FALSE)
}
\arguments{
\item{values}{Colour values}

\item{discrete}{Is the palette being fitted to discrete or continuous data?
The default, FALSE, interpolates a continuous colour palette from the
supplied colour values, and TRUE selects \code{n} evenly spaced colours from the
supplied colour values.}
}
\description{
Create a palette, from the supplied colour values, either selecting \code{n}
evenly spaced colours from the colour values, if \code{discrete = TRUE}, or an
interpolated palette over the full length of the colour values.
}
