% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sequential.R
\name{pal_sequential}
\alias{pal_sequential}
\title{Sequential Colour Palettes}
\usage{
pal_sequential(palette = c("blue_mono", "blue_green"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there are two sequential colour palettes available:
\code{"blue_mono"} (a 9-colour blue monochromatic palette) and
\code{"blue_green"} (a 9-colour blue-to-green sequential palette that starts
with SCW Dark Blue).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Return the SCW sequential colour palettes.
}
\examples{
scales::show_col(pal_sequential("blue_mono")(9))
scales::show_col(pal_sequential("blue_green", alpha = 0.7)(9))
}
