% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Complete list of all SCW colour palettes}
\format{
A list of 5 lists.
}
\usage{
palettes
}
\description{
A list of all SCW colour palettes. Structured as a list of lists with the
first level being each palette type (NHS, SCW, Sequential, Diverging, and
Qualitative) and the second level being the palettes available for each
palette type.
}
\keyword{datasets}
