% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_images.R
\name{save_images}
\alias{save_images}
\title{save_images}
\usage{
save_images(x)
}
\arguments{
\item{x}{Vector relating to index column in the output of get_images(). For
example if x equals c(1,2,4) then this will save images in the output of
get_images() with index 1,2 and 4 to the current working directory.}
}
\description{
Save images from the get_images() function into your current working
directory based on the tabular index.
}
