% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_scw.R
\name{scale_colour_scw}
\alias{scale_colour_scw}
\title{Colour scale constructor for SCW theme colours}
\usage{
scale_colour_scw(palette = "seq", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in SCW theme}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_colour_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour scale constructor for SCW theme colours
}
