% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-diverging.R
\name{scale_colour_diverging}
\alias{scale_colour_diverging}
\alias{scale_color_diverging}
\alias{scale_fill_diverging}
\title{SCW Diverging Colour Palettes}
\usage{
scale_colour_diverging(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_green", "blue_yellow_red", "blue_red")
)

scale_color_diverging(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_green", "blue_yellow_red", "blue_red")
)

scale_fill_diverging(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_green", "blue_yellow_red", "blue_red")
)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{The default, FALSE, generates the colour palette as defined,
while TRUE generates the colour palette in reverse order.}

\item{discrete}{data Whether data the palette is being applied to is discrete
or continuous.}

\item{palette}{Currently there are two diverging colour palettes available:
\code{"blue_green"} (a 9-colour blue-grey-green palette using NHS Blue, NHS
Pale Grey, and NHS Aqua Green), \code{blue_yellow_red} (a 9-colour
blue-yellow-red palette using NHS Blue and NHS Pale Grey), and
\code{blue_red} (a 9-colour blue-to-red palette uses NHS Blue and NHS Pale
Grey).}
}
\description{
See \code{\link{pal_diverging}} for details.
}
