% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_scw.R
\name{scale_fill_scw}
\alias{scale_fill_scw}
\title{Fill scale constructor for SCW colours}
\usage{
scale_fill_scw(palette = "qual", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in SCW theme}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for SCW colours
}
