% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-sequential.R
\name{scale_colour_sequential}
\alias{scale_colour_sequential}
\alias{scale_color_sequential}
\alias{scale_fill_sequential}
\title{SCW Sequential Colour Palettes}
\usage{
scale_colour_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blues", "dark_blues", "blue_grey", "blue_green_yellow")
)

scale_color_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blues", "dark_blues", "blue_grey", "blue_green_yellow")
)

scale_fill_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blues", "dark_blues", "blue_grey", "blue_green_yellow")
)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{The default, FALSE, generates the colour palette as defined,
while TRUE generates the colour palette in reverse order.}

\item{discrete}{data Whether data the palette is being applied to is discrete
or continuous.}

\item{palette}{Currently there are four sequential colour palettes available:
\code{"blues"} (a 9-colour single-hue NHS Blue palette), \code{"dark_blues"}
(a 9-colour single-hue SCW Dark Blue palette), \code{"blue_grey"} (a 9-colour
multi-hue palette that uses SCW Dark Blue and NHS Pale Grey), and
\code{"blue_green_yellow"} (a 9-colour multi-hue palette that is generated
using NHS Blue, NHS Aqua Green, and NHS Warm Yellow).}
}
\description{
See \code{\link{pal_sequential}} for details.
}
