% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scw_colours.R
\name{scw_palette}
\alias{scw_palette}
\title{Return function to interpolate an SCW colour palette}
\usage{
scw_palette(palette = "scw", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in scw_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an SCW colour palette
}
