% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCW_theme.R
\name{scw_theme}
\alias{scw_theme}
\title{Add SCW theme to ggplot chart}
\usage{
scw_theme()
}
\description{
This function allows you to add the SCW theme to your ggplot graphics.
}
\examples{
line <- ggplot(line_df, aes(x = year, y = lifeExp)) +
geom_line(colour = "#007f7f", size = 1) +
geom_hline(yintercept = 0, size = 1, colour="#333333") +
scw_theme()
}
\keyword{ggplot}
\keyword{theme}
