% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scw_theme.R
\name{theme_scw}
\alias{theme_scw}
\title{Add SCW theme to ggplot chart}
\usage{
theme_scw(base_size = 15, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts}

\item{base_family}{base font family}
}
\description{
This function allows you to add the SCW theme to your ggplot graphics.
}
\examples{
library(ggplot2)

ggplot(mpg, aes(displ, cty)) +
  geom_point() +
  theme_scw()

}
\keyword{ggplot}
\keyword{theme}
