% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_diverging.R
\name{pal_diverging}
\alias{pal_diverging}
\title{Diverging Colour Palettes}
\usage{
pal_diverging(
  palette = c("blue_green", "blue_yellow_red", "blue_red"),
  alpha = 1
)
}
\arguments{
\item{palette}{Palette type.
Currently there are three diverging colour palettes available:
\code{"blue_green"} (a 9-colour blue-grey-green palette using NHS Blue, NHS
Pale Grey, and NHS Aqua Green), \code{blue_yellow_red} (a 9-colour
blue-yellow-red palette using NHS Blue and NHS Pale Grey), and
\code{blue_red} (a 9-colour blue-to-red palette uses NHS Blue and NHS Pale
Grey).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Return the SCW diverging colour palettes.
}
\examples{
scales::show_col(pal_diverging("blue_green")(9))
scales::show_col(pal_diverging("blue_yellow_red")(9))
scales::show_col(pal_diverging("blue_red", alpha = 0.7)(9))
}
