% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-qualitative.R
\name{pal_qualitative}
\alias{pal_qualitative}
\title{Qualitative Colour Palettes}
\usage{
pal_qualitative(palette = c("nhs", "scw", "scw_warm"), alpha = 1)
}
\arguments{
\item{palette}{Currently there are three qualitative colour palettes
available: \code{"scw"} (a 5-colour palette that includes SCW Dark Blue, NHS
Aqua Green, NHS Blue, and NHS Light Green), \code{"nhs"} (a 5-colour palette
that includes NHS Blue, NHS Light Blue, NHS Mid Grey, and NHS Dark Blue), and
\code{scw_warm} (a 6-colour palette that includes NHS Blue, NHS Light Blue,
SCW Dark Blue, NHS Orange, and NHS Mid Grey).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Qualitative (sometimes called categorical) colour palettes are used to encode
categorical information that has no meaningful order. For example, car
manufacturers or sports teams, without context that specifically indicates
some natural order, would be unordered categories. Qualitative colour
palettes should include colours that all have similar "perceptual weights",
meaning that each colour is clearly distinguishable from each other, but that
no colour is significantly more distinct than the rest of the palette.
}
\examples{
scales::show_col(pal_qualitative(palette = "nhs")(5))
scales::show_col(pal_qualitative(palette = "scw")(5))
scales::show_col(pal_qualitative(palette = "scw_warm")(5))
}
