% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-qualitative.R
\name{pal_qualitative}
\alias{pal_qualitative}
\title{Qualitative Colour Palettes}
\usage{
pal_qualitative(palette = c("default"), alpha = 1)
}
\arguments{
\item{palette}{Palette type.
Currently there is only one qualitative colour palette available:
\code{"default"} (a 5-colour qualitative palette that includes NHS Blue,
NHS Warm Yellow, and SCW Dark Blue).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}
}
\description{
Return the SCW qualitative colour palettes.
}
\examples{
scales::show_col(pal_qualitative(palette = "default")(5))
}
