% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_diverging.R
\name{scale_colour_diverging}
\alias{scale_colour_diverging}
\alias{scale_color_diverging}
\alias{scale_fill_diverging}
\title{SCW diverging Colour Palettes}
\usage{
scale_colour_diverging(
  palette = c("blue_green", "blue_yellow_red", "blue_red"),
  discrete = FALSE,
  alpha = 1,
  ...
)

scale_color_diverging(
  palette = c("blue_green", "blue_yellow_red", "blue_red"),
  discrete = FALSE,
  alpha = 1,
  ...
)

scale_fill_diverging(
  palette = c("blue_green", "blue_yellow_red", "blue_red"),
  alpha = 1,
  discrete = TRUE,
  ...
)
}
\arguments{
\item{palette}{Palette type.
Currently there are three diverging colour palettes available:
\code{"blue_green"} (a 9-colour blue-grey-green palette using NHS Blue, NHS
Pale Grey, and NHS Aqua Green), \code{blue_yellow_red} (a 9-colour
blue-yellow-red palette using NHS Blue and NHS Pale Grey), and
\code{blue_red} (a 9-colour blue-to-red palette uses NHS Blue and NHS Pale
Grey).}

\item{discrete}{data Whether data the palette is being applied to is discrete
or continuous.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_diverging}} for details.
}
