% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-qualitative.R
\name{scale_colour_qualitative}
\alias{scale_colour_qualitative}
\alias{scale_color_qualitative}
\alias{scale_fill_qualitative}
\title{SCW Qualitative Colour Palettes}
\usage{
scale_colour_qualitative(palette = c("default", "nhs", "scw"), alpha = 1, ...)

scale_color_qualitative(palette = c("default", "nhs", "scw"), alpha = 1, ...)

scale_fill_qualitative(palette = c("default", "nhs", "scw"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there are three qualitative colour palettes available:
\code{"default"} (a 6-colour saturated qualitative palette that includes NHS
Blue, NHS Warm Yellow, and SCW Dark Blue), \code{"nhs"} (a 6-colour
qualitative palette that includes NHS Blue, NHS Warm Yellow, NHS Pink, NHS
Purple, and NHS Light Blue), and \code{scw} (a 6-colour qualitative palette
that is made up of entirely NHS/SCW colours).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_qualitative}} for details.
}
