% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-qualitative.R, R/zxx.R
\name{scale_colour_qualitative}
\alias{scale_colour_qualitative}
\alias{scale_fill_qualitative}
\alias{scale_color_qualitative}
\title{SCW Qualitative Colour Palettes}
\usage{
scale_colour_qualitative(palette = c("nhs", "scw", "warm"), alpha = 1, ...)

scale_fill_qualitative(palette = c("nhs", "scw", "warm"), alpha = 1, ...)
}
\arguments{
\item{palette}{Currently there are three qualitative colour palettes
available: \code{"scw"} (a 5-colour palette that includes NHS Blue, NHS
Light Blue, and SCW Dark Blue), \code{"nhs"} (a 5-colour palette that
includes NHS Dark Blue, NHS Blue, NHS Light Blue, and NHS Mid Grey), and
\code{warm} (a 6-colour palette that includes NHS Blue, NHS Warm Yellow,
NHS Mid Grey, and SCW Dark Blue).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_qualitative}} for details.
}
