% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_qualitative.R
\name{scale_colour_qualitative}
\alias{scale_colour_qualitative}
\alias{scale_color_qualitative}
\alias{scale_fill_qualitative}
\title{SCW Qualitative Colour Palettes}
\usage{
scale_colour_qualitative(palette = c("default"), alpha = 1, ...)

scale_color_qualitative(palette = c("default"), alpha = 1, ...)

scale_fill_qualitative(palette = c("default"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there is only one qualitative colour palette available:
\code{"default"} (a 5-colour qualitative palette that includes NHS Blue,
NHS Warm Yellow, and SCW Dark Blue).}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_qualitative}} for details.
}
