% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sequential.R
\name{scale_colour_sequential}
\alias{scale_colour_sequential}
\alias{scale_color_sequential}
\alias{scale_fill_sequential}
\title{SCW Sequential Colour Palettes}
\usage{
scale_colour_sequential(
  palette = c("blue_mono", "blue_green"),
  discrete = FALSE,
  alpha = 1,
  ...
)

scale_color_sequential(
  palette = c("blue_mono", "blue_green"),
  discrete = FALSE,
  alpha = 1,
  ...
)

scale_fill_sequential(palette = c("blue_mono", "blue_green"), alpha = 1, ...)
}
\arguments{
\item{palette}{Palette type.
Currently there are two sequential colour palettes available:
\code{"blue_mono"} (a 9-colour blue monochromatic palette) and
\code{"blue_green"} (a 9-colour blue-to-green sequential palette that starts
with SCW Dark Blue).}

\item{discrete}{data Whether data the palette is being applied to is discrete
or continuous.}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
See \code{\link{pal_sequential}} for details.
}
