% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-sequential.R
\name{scale_colour_sequential}
\alias{scale_colour_sequential}
\alias{scale_color_sequential}
\alias{scale_fill_sequential}
\title{SCW Sequential Colour Palettes}
\usage{
scale_colour_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_mono", "blue_green")
)

scale_color_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_mono", "blue_green")
)

scale_fill_sequential(
  ...,
  alpha = 1,
  reverse = TRUE,
  discrete = FALSE,
  palette = c("blue_mono", "blue_green")
)
}
\arguments{
\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}

\item{alpha}{Transparency level, a real number in (0, 1].
See \code{alpha} in \code{\link[grDevices]{rgb}} for details.}

\item{reverse}{The default, FALSE, generates the colour palette as defined,
while TRUE generates the colour palette in reverse order.}

\item{discrete}{data Whether data the palette is being applied to is discrete
or continuous.}

\item{palette}{Currently there are two sequential colour palettes available:
\code{"blue_mono"} (a 9-colour blue monochromatic palette) and
\code{"blue_green"} (a 9-colour blue-to-green sequential palette that starts
with SCW Dark Blue).}
}
\description{
See \code{\link{pal_sequential}} for details.
}
